support sorting more than 4 billions element

+++ parallel_prac/2017autumn/task2/Point.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stdint.h>
 #include <mpi.h>
 
 class Point {
@@ -12,7 +13,7 @@ public:
     static MPI::Datatype datatype();
 
     float x, y;
-    int index;
+    int64_t index;
 private:
     static sort_way sort_by;
     static MPI::Datatype _datatype;
